+++ src-docs/src/views/tables/in_memory/in_memory_search_callback.js
@@ -44,7 +44,7 @@ export class Table extends React.Component {
     };
   }
 
-  onQueryChange = query => {
+  onQueryChange = (query, onQueryComplete) => {
     clearTimeout(debounceTimeoutId);
     clearTimeout(requestTimeoutId);
 
@@ -64,6 +64,8 @@ export class Table extends React.Component {
           isLoading: false,
           items,
         });
+
+        onQueryComplete();
       }, 1000);
     }, 300);
   };