+++ src-docs/src/views/tables/in_memory/in_memory_search_callback_section.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { Fragment } from 'react';
 import {
   EuiCode
 } from '../../../../../src/components';
@@ -23,12 +23,17 @@ export const searchCallbackSection = {
     }
   ],
   text: (
-    <div>
+    <Fragment>
       <p>
         The example shows how to configure <EuiCode>EuiInMemoryTable</EuiCode> to display a search bar
         and intercept the search value when it changes so you can perform your own search logic.
       </p>
-    </div>
+
+      <p>
+        Note that when the search has returned results, you&rsquo;ll need to call the <EuiCode>onQueryComplete</EuiCode>
+        callback that&rsquo;s provided to the <EuiCode>search.onChange</EuiCode> prop.
+      </p>
+    </Fragment>
   ),
   props: propsInfo,
   demo: <Table/>